Made changes!!
I started this project back up on December 18, 2015. reason for the
break is because I enrolled in another unity course to gain more
knowledge of unity.
Removed the 25 questions scenes from the build, all of the questions
that belong to a gameboard will be on a penal that will display each
question, got the buttons to set to not active after being pressed, and
the gameboard canvas will not be active when a question is displayed.
All of the main functions of the game is finished except for the score,
and I got rid of the background for the gameboard for now.