Finished the Gameboard scene UI!!
I Made a mistake I started the project back up on  December 19, 2015. I
have the Gameboard scene UI done. I have the categories at the top of
the scene; Also the score, which I have named knowledge points, and also
have the score displayed after the name . 25  question buttons are in
the scene and some of the buttons are colored with Sigma Chi colors and
they change colors when hovered over and pressed on.