A very long time consuming process complete
I took a break from 12/252015 to 1/20/2016 I was sick for a part of that
time. I made a simple scoring system. Then I made changes to my action
script, instead of a reward and lose methods for general purpose.  Now
there is a method for each reward and lose condition. I made it as
simple as possible!! during the long process I found some errors in two
of my questions, and made the changes on the question document as well.
I also Found a copy of an answer for one of my questions fixed that as
well on the game side. Lastly I connected the question buttons to the
questions. Phew!! that took hours!!! It was fun!!!