Finally almost done with this small game!!!!
Before I got started programming today I watched Meet The Robinsons made
by Disney. The movie is a heart- felt movie about not giving up and just
moving forward. I will recommend this movie to anyone. I only worked on
this game one day this week, and it just happened to be today. I
finished the transitions to the Endgane Panel from the GameBoard, in
fact the EndGame Panel places itself on top of the GameBoard after the
last question Button is pressed and the Question Panel is inactive. The
categories text goes inactive as well and I removed the score from the
Endgame Panel. After that was done I designed the Main Menu Scene better
then it was. I Also made a donations button and Credits button. Then
made a PayPay account for me to accept donations  and made a added code
to my LoadScene Script to navigate to the donations page when the
donation button is clicked. Finally I made made the credits Scene and
added a back button to it to get back to the Main Menu Scene. Oh yeah, l
Linked the project to UnityAds properly to get rid of an error message.